requirements.txt
numpy
pytorch
torchvision
pandas
tqdm
